+++ NEWS.md
@@ -1,3 +1,5 @@
+* Added the function `add_kaggle` to quickly write bibtex for kaggle kernel citation.
+
 # add2bibtex 0.2.0
 
 * Added the current date feature in the functions.